Takes an integer and converts it into a token.
Takes a token and converts it into an integer.
Determining the next token from two consecutive tokens is NP-Hard.
The alphabet and generator is determined at runtime and thus will generate new tokens for the same numbers on restart.
Can create up to MAX_ROOMS urls. (This value can be changed if needed).